- Modificada en el Makefile la regla de la clase Classifier para incluir la API Boost. - Añadido en la función main la comprobación del número de argumentos pasados al programa. - Cambiados todos los tipos de datos float a double. - Solucionado un bug en el método Segmenter::findLineMarkers() que no tenía en cuenta que la última fila de pixels de un muestra fuera de tinta y no de fondo. - Modificada la condición para saber si es necesario unir a posteriori dos líneas detectadas en el método Segmenter::findLineMarkers(). - Añadidos mensajes de depuración a lo largo de la implementación de las clases Segmenter, Recognizer y Classifier. - Modificado el método Recognizer::obtainText() para añadir la etapa de clasificación. - Solucionado un bug en el método Shape::normalizeCoordinates() que asignaba a la coordenada Y el resultado de la normalización de la columna X. - Añadido el método DataSet::countCategories(). - Añadida una restricción en el constructor de la clase DataSet para evitar que se utilicen data sets con más de 7 características. - Implementados los métodos Classifier::buildFeatureVector() y Classifier::matchShape(). Además, en la clase Classifier se han añadido las funciones locales lambda() y computeCentralMoment().

+++ trunk/inc/Classifier.hpp
@@ -75,7 +75,7 @@ public:
 	/// @author Eliezer Talón (elitalon@gmail.com)
 	/// @date 2008-11-07
 	///
-	const float& matchingShapeTime ();
+	const double& matchingShapeTime ();
 
 
 private:
@@ -84,7 +84,7 @@ private:
 	
 	FeatureVector	sample_;			///< A sample that represents a shape.
 	
-	float			matchingShapeTime_;	///< Elapsed time while matching a shape into a character.
+	double			matchingShapeTime_;	///< Elapsed time while matching a shape into a character.
 	
 	
 	///
@@ -116,7 +116,7 @@ private:
 
 
 
-inline const float& Classifier::matchingShapeTime ()
+inline const double& Classifier::matchingShapeTime ()
 {
 	return matchingShapeTime_;
 };
